requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
openai>=1.30.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.7.0
streamlit>=1.37.0
requests>=2.32.0
